Attached users to database, combined js files, no more browserify
Pretty much had to shift everything around to get this to work but
now users are stored on realtime database and can be updated from the
game - currently only win count and loss count are implementing this.
Resolves #12

Big issue with turns - on game start nothing happens until a space is
clicked when 'ActivePlayer' is set to be current player. If AI is set,
this works fine - probably a load order issue - Also if 'ActivePlayer'
wins, he gets 2 turns to start, this doesn't happen if AI wins